pyproject.toml
[project]
name = "littledsm"
version = "0.1.0"
description = "FastAPI development setup for littleDSM"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.0,<1.0.0",
    "python-jose[cryptography]>=3.5.0,<4.0.0",
    "sqlalchemy>=2.0.41,<3.0.0",
    "uvicorn[standard]>=0.35.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.28.0,<1.0.0",
    "pytest>=8.4.0,<9.0.0",
    "pytest-asyncio>=1.1.0,<2.0.0",
    "ruff>=0.12.0,<1.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
